+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## CHANGELOG
 
-### 4.0.0b14 (21-Apr-2026)
+### 4.0.0b15 (23-Apr-2026)
 - Added JSON health reporting (with optional Splunk HTTP Event Collector (HEC) delivery)
 - Added a stand-alone swiftDialog Inspect Mode-flavored report (i.e., `inspectSummaryPreset="on"`), plus cached replay (i.e., `inspectReplayMaximumAgeSeconds`) for `Self Service` runs
 - Refactored `checkElectronCornerMask` to reduce execution time
@@ -12,6 +12,8 @@
 - Added "Next Steps" to Inspect Mode-flavored report
 - Added `checkWiFiStrength()`; thanks, @kgolden-code!
 - Removed `displayFailureNotification()` in favor of the Inspect Mode-flavored report
+- Refactored `Silent` when used with `splunkOperationMode=production` to suppress non-Splunk console output while still logging fully to `scriptLog`, and to return success only after Splunk report generation plus HEC delivery succeed
+- Refactored `updateComputerInventory()` to log and skip `jamf recon` during `Silent` runs when `splunkOperationMode=production`
 
 ### 3.0.0 (23-Feb-2026)
 **First (attempt at a) MDM-agnostic release**